requirements.txt
# Основные зависимости
fastapi>=0.104.1
uvicorn>=0.24.0
sqlalchemy>=2.0.23
alembic>=1.12.1
asyncpg>=0.29.0
pydantic>=2.5.1
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
loguru>=0.7.2
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
websockets>=12.0 
typer

# Тестовые зависимости
typer[all]>=0.9.0
passlib[bcrypt]>=1.7.4
pytest>=7.4.3
pytest-asyncio>=0.21.1
httpx>=0.25.1

# Добавленные зависимости
email-validator==2.1.0
pydantic[email]>=2.0.0

